backend/requirementsé.txt
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.1
decorator==5.2.1
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==5.1.6
django-allauth==0.57.0
django-cors-headers==4.7.0
django-filter==25.1
django-imagefield==0.22.0
django-phonenumber-field==8.0.0
django-rest-swagger==2.2.0
django-spaghetti-and-meatballs==0.4.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docopt==0.6.2
drf-yasg==1.21.9
executing==2.2.0
fastjsonschema==2.21.1
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.68.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
ipython==8.12.3
itypes==1.2.0
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.2
msgpack==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ngrok==1.4.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
phonenumbers==9.0.0
phonenumberslite==8.13.55
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
redis==5.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
rsa==4.9
simplejson==3.20.1
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
